Increase covariance of odom 'z' twist
This change was found to reduce the rotational offset noticed in the lidar scans when the robot rotates in a circle.  As a result, the scans and localization seem to be more stable when rotating.  This probably makes sense since for a skid steer robot, the quality of the odom twist info is poor since by design the wheels will slip/skid.  Increasing the odom covariance then gives more priority to the much better IMU measurement of rotation about z.